+++ src/com/intellij/idea/plugin/hybris/system/bean/meta/BSMetaModelAccess.kt
@@ -17,7 +17,6 @@
  */
 package com.intellij.idea.plugin.hybris.system.bean.meta
 
-import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils
 import com.intellij.idea.plugin.hybris.system.bean.meta.impl.BSMetaModelNameProvider
 import com.intellij.idea.plugin.hybris.system.bean.meta.model.BSGlobalMetaBean
 import com.intellij.idea.plugin.hybris.system.bean.meta.model.BSGlobalMetaClassifier
@@ -25,29 +24,23 @@ import com.intellij.idea.plugin.hybris.system.bean.meta.model.BSGlobalMetaEnum
 import com.intellij.idea.plugin.hybris.system.bean.meta.model.BSMetaType
 import com.intellij.idea.plugin.hybris.system.bean.model.Bean
 import com.intellij.idea.plugin.hybris.system.bean.model.Enum
-import com.intellij.openapi.application.ApplicationManager
-import com.intellij.openapi.application.ReadAction
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.progress.ProcessCanceledException
-import com.intellij.openapi.progress.ProgressIndicator
-import com.intellij.openapi.progress.ProgressManager
-import com.intellij.openapi.progress.Task
-import com.intellij.openapi.progress.impl.BackgroundableProcessIndicator
 import com.intellij.openapi.project.DumbService
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.util.Computable
 import com.intellij.openapi.util.Key
 import com.intellij.openapi.util.ModificationTracker
+import com.intellij.platform.ide.progress.withBackgroundProgress
+import com.intellij.platform.util.progress.reportProgress
 import com.intellij.psi.PsiFile
 import com.intellij.psi.util.CachedValue
 import com.intellij.psi.util.CachedValueProvider
 import com.intellij.psi.util.CachedValuesManager
 import com.intellij.util.messages.Topic
-import java.util.*
-import java.util.concurrent.Semaphore
+import kotlinx.coroutines.*
 
 @Service(Service.Level.PROJECT)
-class BSMetaModelAccess(private val myProject: Project) {
+class BSMetaModelAccess(private val project: Project, private val coroutineScope: CoroutineScope) {
 
     companion object {
         private val SINGLE_MODEL_CACHE_KEY = Key.create<CachedValue<BSMetaModel>>("SINGLE_BEAN_SYSTEM_MODEL_CACHE")
@@ -57,67 +50,65 @@ class BSMetaModelAccess(private val myProject: Project) {
     }
 
     private val myGlobalMetaModel = BSGlobalMetaModel()
-    private val myMessageBus = myProject.messageBus
+    private val myMessageBus = project.messageBus
 
     @Volatile
     private var building: Boolean = false
 
     @Volatile
     private var initialized: Boolean = false
-    private val semaphore = Semaphore(1)
 
-    private val myGlobalMetaModelCache = CachedValuesManager.getManager(myProject).createCachedValue(
+    private val myGlobalMetaModelCache = CachedValuesManager.getManager(project).createCachedValue(
         {
-            val localMetaModels = BSMetaModelCollector.getInstance(myProject).collectDependencies()
-                .filter { obj: PsiFile? -> Objects.nonNull(obj) }
-                .map { psiFile: PsiFile -> retrieveSingleMetaModelPerFile(psiFile) }
-                .map { obj: CachedValue<BSMetaModel> -> obj.value }
-                .sortedBy { !it.custom }
+            val localMetaModels = runBlocking {
+                withBackgroundProgress(project, "Re-building Bean System...", true) {
+                    val collectedDependencies = BSMetaModelCollector.getInstance(project).collectDependencies()
 
+                    val localMetaModels = reportProgress(collectedDependencies.size) { progressReporter ->
+                        collectedDependencies
+                            .map {
+                                progressReporter.sizedStep(1, "Processing: ${it.name}...") {
+                                    this.async {
+                                        retrieveSingleMetaModelPerFile(it)
+                                    }
+                                }
+                            }
+                            .awaitAll()
+                            .sortedBy { !it.custom }
+                    }
+
+                    BSMetaModelMerger.merge(myGlobalMetaModel, localMetaModels)
+
+                    localMetaModels
+                }
+            }
             val dependencies = localMetaModels
                 .map { it.psiFile }
                 .toTypedArray()
-            BSMetaModelMerger.merge(myGlobalMetaModel, localMetaModels)
 
             CachedValueProvider.Result.create(myGlobalMetaModel, dependencies.ifEmpty { ModificationTracker.EVER_CHANGED })
         }, false
     )
 
-    private val task = object : Task.Backgroundable(myProject, HybrisI18NBundleUtils.message("hybris.bs.access.progress.title.building")) {
-        override fun run(indicator: ProgressIndicator) {
-            indicator.text2 = HybrisI18NBundleUtils.message("hybris.bs.access.progress.subTitle.waitingForIndex")
-
-            ReadAction
-                .nonBlocking<Unit> {
-                    val lock = semaphore.tryAcquire()
-
-                    if (lock) {
-                        try {
-                            building = true
-                            val globalMetaModel = myGlobalMetaModelCache.value
-                            building = false
-                            initialized = true
-
-                            myMessageBus.syncPublisher(BSMetaModelAccess.TOPIC).beanSystemChanged(globalMetaModel)
-                        } finally {
-                            semaphore.release()
-                        }
-                    }
-                }
-                .inSmartMode(project)
-                .executeSynchronously()
-        }
-    }
-
     fun isInitialized() = initialized
 
     fun initMetaModel() {
         building = true
-        ProgressManager.getInstance().runProcessWithProgressAsynchronously(task, BackgroundableProcessIndicator(task))
+
+        coroutineScope
+            .launch(Dispatchers.IO) {
+                myGlobalMetaModelCache.value
+            }
+            .invokeOnCompletion {
+                building = false
+                initialized = true
+
+                myMessageBus.syncPublisher(TOPIC).beanSystemChanged(myGlobalMetaModel)
+            }
     }
 
     fun getMetaModel(): BSGlobalMetaModel {
-        if (building || !initialized || DumbService.isDumb(myProject)) throw ProcessCanceledException()
+        if (building || !initialized || DumbService.isDumb(project)) throw ProcessCanceledException()
 
         if (myGlobalMetaModelCache.hasUpToDateValue()) {
             return myGlobalMetaModelCache.value
@@ -167,23 +158,15 @@ class BSMetaModelAccess(private val myProject: Project) {
     private fun <T : BSGlobalMetaClassifier<*>> findMetaByName(metaType: BSMetaType, name: String?): T? =
         getMetaModel().getMetaType<T>(metaType)[name]
 
-    private fun retrieveSingleMetaModelPerFile(psiFile: PsiFile): CachedValue<BSMetaModel> {
-        return Optional.ofNullable(psiFile.getUserData(SINGLE_MODEL_CACHE_KEY))
-            .orElseGet {
-                val cachedValue = createSingleMetaModelCachedValue(myProject, psiFile)
-                psiFile.putUserData(SINGLE_MODEL_CACHE_KEY, cachedValue)
-                cachedValue
+    private fun retrieveSingleMetaModelPerFile(psiFile: PsiFile): BSMetaModel = CachedValuesManager.getManager(project).getCachedValue(
+        psiFile, SINGLE_MODEL_CACHE_KEY,
+        {
+            val value = runBlocking {
+                BSMetaModelProcessor.getInstance(project).process(this, psiFile)
             }
-    }
 
-    private fun createSingleMetaModelCachedValue(project: Project, psiFile: PsiFile): CachedValue<BSMetaModel> {
-        return CachedValuesManager.getManager(project).createCachedValue(
-            {
-                ApplicationManager.getApplication().runReadAction(
-                    Computable {
-                        CachedValueProvider.Result.create(BSMetaModelProcessor.getInstance(myProject).process(psiFile), psiFile)
-                    } as Computable<CachedValueProvider.Result<BSMetaModel>>)
-            }, false
-        )
-    }
+            CachedValueProvider.Result.create(value, psiFile)
+        },
+        false
+    )
 }