+++ src/com/intellij/idea/plugin/hybris/system/bean/meta/BSMetaModelCollector.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -17,11 +17,10 @@
  */
 package com.intellij.idea.plugin.hybris.system.bean.meta
 
-import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils
 import com.intellij.idea.plugin.hybris.system.bean.BSUtils
 import com.intellij.idea.plugin.hybris.system.bean.model.Beans
+import com.intellij.openapi.application.readAction
 import com.intellij.openapi.components.Service
-import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
 import com.intellij.psi.search.ProjectScope
@@ -41,10 +40,8 @@ class BSMetaModelCollector(private val myProject: Project) {
 
     private val myDomManager: DomManager = DomManager.getDomManager(myProject)
 
-    fun collectDependencies(): Set<PsiFile> {
+    suspend fun collectDependencies(): Set<PsiFile> = readAction {
         val files = HashSet<PsiFile>()
-
-        ProgressManager.getInstance().progressIndicator.text2 = HybrisI18NBundleUtils.message("hybris.bs.access.progress.subTitle.collectingDependencies")
 
         StubIndex.getInstance().processElements(
             DomElementClassIndex.KEY,
@@ -65,8 +62,7 @@ class BSMetaModelCollector(private val myProject: Project) {
             }
         )
 
-        ProgressManager.getInstance().progressIndicator.text2 = HybrisI18NBundleUtils.message("hybris.bs.access.progress.subTitle.collectedDependencies", files.size)
-
-        return Collections.unmodifiableSet(files)
+        files
     }
+        .let { Collections.unmodifiableSet(it) }
 }