+++ src/com/intellij/idea/plugin/hybris/system/bean/meta/BSMetaModelProcessor.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -17,16 +17,16 @@
  */
 package com.intellij.idea.plugin.hybris.system.bean.meta
 
-import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils
 import com.intellij.idea.plugin.hybris.system.bean.BSUtils
 import com.intellij.idea.plugin.hybris.system.bean.meta.impl.BSMetaModelBuilder
 import com.intellij.idea.plugin.hybris.system.bean.model.Beans
+import com.intellij.openapi.application.readAction
 import com.intellij.openapi.components.Service
-import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
 import com.intellij.psi.xml.XmlFile
 import com.intellij.util.xml.DomManager
+import kotlinx.coroutines.*
 
 @Service(Service.Level.PROJECT)
 class BSMetaModelProcessor(myProject: Project) {
@@ -34,22 +34,29 @@ class BSMetaModelProcessor(myProject: Project) {
     companion object {
         fun getInstance(project: Project): BSMetaModelProcessor = project.getService(BSMetaModelProcessor::class.java)
     }
+
     private val myDomManager: DomManager = DomManager.getDomManager(myProject)
 
-    fun process(psiFile: PsiFile): BSMetaModel? {
-        psiFile.virtualFile ?: return null
-        val module = BSUtils.getModuleForFile(psiFile) ?: return null
+    suspend fun process(coroutineScope: CoroutineScope, psiFile: PsiFile): BSMetaModel? = coroutineScope {
+        psiFile.virtualFile ?: return@coroutineScope null
+        val module = BSUtils.getModuleForFile(psiFile) ?: return@coroutineScope null
         val custom = BSUtils.isCustomExtensionFile(psiFile)
         val root = myDomManager.getFileElement(psiFile as XmlFile, Beans::class.java)
             ?.rootElement
-            ?: return null
+            ?: return@coroutineScope null
 
-        ProgressManager.getInstance().progressIndicator.text2 = HybrisI18NBundleUtils.message("hybris.bs.access.progress.subTitle.processing", psiFile.name)
+        val builder = BSMetaModelBuilder(module, psiFile, custom)
 
-        return BSMetaModelBuilder(module, psiFile, custom)
-            .withEnumTypes(root.enums)
-            .withBeanTypes(root.beans)
-            .withEventTypes(root.beans)
-            .build()
+        val operations = listOf(
+            coroutineScope.async { readAction { builder.withEnumTypes(root.enums) } },
+            coroutineScope.async { readAction { builder.withBeanTypes(root.beans) } },
+            coroutineScope.async { readAction { builder.withEventTypes(root.beans) } },
+        )
+
+        withContext(Dispatchers.IO) {
+            operations.awaitAll()
+        }
+
+        builder.build()
     }
 }